Rename things to make the UnconfirmedMessageHandler independent
