+++ pkg/retry/unconfirmed_message_handler.go
@@ -1,18 +1,21 @@
-package compliance
+package retry
 
 import (
 	"context"
 	"fmt"
 	"time"
+
+	"github.com/stackrox/rox/pkg/logging"
 )
 
-var _ unconfirmedMessageHandler = (*NodeScanResend)(nil)
+var (
+	log                 = logging.LoggerForModule()
+	defaultBaseInterval = 1 * time.Minute
+)
 
-var defaultBaseInterval = 1 * time.Minute
-
-// NodeScanResend handles resending node scans based on ACK messages from Central
-// Assumption: Time to receive an ACK is generally much shorter than ROX_NODE_SCANNING_INTERVAL.
-type NodeScanResend struct {
+// UnconfirmedMessageHandlerImpl informs the caller when a resending should happen based on ACK messsage.
+// Assumption: Time to receive an ACK is generally much shorter than the interval between sending consequtive messages.
+type UnconfirmedMessageHandlerImpl struct {
 	// baseInterval defines the delay after which we resend a message
 	baseInterval time.Duration
 	// resendInterval is a multiply of the baseInterval and defines how much time should pass before the next retry
@@ -27,10 +30,10 @@ type NodeScanResend struct {
 	ctx context.Context
 }
 
-// NewNodeScanResend returns a new running NodeScanResend.
+// NewUnconfirmedMessageHandler returns a new running UnconfirmedMessageHandlerImpl.
 // It can be stopped by canceling the context.
-func NewNodeScanResend(ctx context.Context, resendInterval time.Duration) *NodeScanResend {
-	nsr := &NodeScanResend{
+func NewUnconfirmedMessageHandler(ctx context.Context, resendInterval time.Duration) *UnconfirmedMessageHandlerImpl {
+	nsr := &UnconfirmedMessageHandlerImpl{
 		baseInterval:   resendInterval,
 		resendInterval: resendInterval,
 		ticker:         time.NewTicker(resendInterval),
@@ -44,19 +47,19 @@ func NewNodeScanResend(ctx context.Context, resendInterval time.Duration) *NodeS
 }
 
 // RetryCommand returns a channel that will produce a message when sending should be retried
-func (s *NodeScanResend) RetryCommand() <-chan struct{} {
+func (s *UnconfirmedMessageHandlerImpl) RetryCommand() <-chan struct{} {
 	return s.ch
 }
 
 // run starts the ticker, so that resending can happen after defined time without ACK passes
-func (s *NodeScanResend) run() {
+func (s *UnconfirmedMessageHandlerImpl) run() {
 	go func() {
 		defer close(s.ch)
 		for {
 			select {
 			case <-s.ticker.C:
 				s.retryLater()
-				log.Infof("Resending node scan, retry %d (next retry in %s)", s.retry, s.resendInterval)
+				log.Infof("Suggesting to resend, retry %d (next retry in %s)", s.retry, s.resendInterval)
 				s.ch <- struct{}{}
 			case <-s.ctx.Done():
 				return
@@ -65,7 +68,7 @@ func (s *NodeScanResend) run() {
 	}()
 }
 
-func (s *NodeScanResend) retryLater() {
+func (s *UnconfirmedMessageHandlerImpl) retryLater() {
 	s.retry++
 	nextIn := s.retry * int(s.baseInterval.Seconds())
 	next, err := time.ParseDuration(fmt.Sprintf("%ds", nextIn))
@@ -77,17 +80,17 @@ func (s *NodeScanResend) retryLater() {
 	s.ticker.Reset(s.resendInterval)
 }
 
-// ObserveSending should be called when a new node-inventory is sent
-func (s *NodeScanResend) ObserveSending() {
-	log.Infof("Observing node scan being sent. Waiting for an ACK for %s", s.baseInterval.String())
+// ObserveSending should be called when a new message is sent
+func (s *UnconfirmedMessageHandlerImpl) ObserveSending() {
+	log.Infof("Observing message scan being sent. Waiting for an ACK for %s", s.baseInterval.String())
 	s.ticker.Stop()
 	s.retry = 0
 	s.ticker.Reset(s.baseInterval)
 }
 
-// ObserveConfirmation should be called when an ACK for node-inventory is received
-func (s *NodeScanResend) ObserveConfirmation() {
-	log.Info("Node Scan has been acknowledged")
+// ObserveConfirmation should be called when an ACK for the massage is received
+func (s *UnconfirmedMessageHandlerImpl) ObserveConfirmation() {
+	log.Info("Message has been acknowledged")
 	s.ticker.Stop()
 	s.retry = 0
 }
